requirements.txt
flask
redis
python-dotenv
flask-cors
